Implementando vista por modulo
