+++ app/src/main/AndroidManifest.xml
@@ -7,9 +7,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme"
-        >
-
+        android:theme="@style/AppTheme">
         <activity android:name=".Vista.PantallaPpal">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -17,6 +15,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Vista.SeguimientoCognitivo"
+            android:label="@string/title_activity_seguimiento_cognitivo"
+            android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
 
 </manifest>
