+++ app/src/main/java/co/edu/udea/compumovil/gr01_20171/proyectoescuela/Controlador/Controlador.java
@@ -1,0 +1,8 @@
+package co.edu.udea.compumovil.gr01_20171.proyectoescuela.Controlador;
+
+/**
+ * Created by alejandro on 9/03/17.
+ */
+
+public class Controlador {
+}
