+++ app/src/main/res/values/dimens.xml
@@ -3,4 +3,5 @@
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="titulo_1_size">20dp</dimen>
+    <dimen name="fab_margin">16dp</dimen>
 </resources>
