+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="title_activity_seguimiento_cognitivo">SeguimientoCognitivo</string>
 </resources>
